requirements.txt
libvirt-python==3.1.0
